cleanup for easier diff
